vsftpd: fix port number
README says the default port number is 3478, but docker-compose.yml says it's 8004. It has to be fixed. However, I suggest we will change the port used by default to 2121 since in MMB ports starting from 8001 are used for the services the ports of which belong to neither well-known ports nor the registered ports. FTP servers have to listen on well-known port 21. But we don't want our vsftpd to intersect with those which are already installed on the target machine. So, the port number has to be a hint of 21. Let it be 2121.